Remove unused CLI code.
