+++ pyproject.toml
@@ -127,7 +127,7 @@ ignore = [
 ]
 
 [project.scripts]
-datashuttle = "datashuttle.command_line_interface:main"
+datashuttle = "datashuttle.cli_launcher:main"
 
 [tool.codespell]
 skip = '.git,*.pdf,*.svg'
